// frontend/src/backend.d.ts
export enum IngredientSafety {
    safe = "safe",
    conditional = "conditional",
    harmful = "harmful",
    risky = "risky"
}
export enum PriceRange {
    low = "low",
    high = "high",
    medium = "medium"
}
export enum ProductCategory {
    sunscreen = "sunscreen",
    cleanser = "cleanser",
    treatment = "treatment",
    serum = "serum",
    moisturizer = "moisturizer"
}
export enum ProductSuitability {
    caution = "caution",
    suitable = "suitable",
    not_recommended = "not_recommended"
}